+++ src/cljs-lib/src/calva/repl/webview/ui.cljs
@@ -2,12 +2,19 @@
   (:require
    [cljs.reader :as reader]
    ["strip-ansi" :default strip-ansi]
-   ["highlightjs-copy" :as CopyButtonPlugin]))
+   ["highlightjs-copy" :as CopyButtonPlugin]
+   ["highlight.js/lib/core" :as hljs]
+   ["highlight.js/lib/languages/clojure" :as clojure]))
 
 ;; The DOM element where output is written
 (def output-dom-element (js/document.getElementById "output"))
 
-(defonce vscode (js/acquireVsCodeApi))
+(defn ensure-dom-content-loaded
+  "Ensures the DOM is ready before executing the callback"
+  [callback]
+  (if (= "complete" js/document.readyState)
+    (callback)
+    (js/document.addEventListener "DOMContentLoaded" callback #js {:once true})))
 
 (defn throttle-fn
   "Returns a throttled version of the function, which will only be called at most once every `wait`
@@ -70,14 +77,14 @@
     (.. div (appendChild span))
     (.. div (appendChild container-element))
     (.. dom-element (appendChild div))
-    (.. js/window -hljs (highlightElement code-element))
+    (.. hljs (highlightElement code-element))
     (.. dom-element (dispatchEvent (output-appended-event div)))))
 
 (defn append-eval-result
   [^js dom-element {:keys [output]}]
   (let [{:keys [code-element container-element]} (clojure-code-element output)]
     (.. dom-element (appendChild container-element))
-    (.. js/window -hljs (highlightElement code-element))
+    (.. hljs (highlightElement code-element))
     (.. dom-element (dispatchEvent (output-appended-event container-element)))))
 
 (defn create-and-append-stdout-element
@@ -135,81 +142,31 @@
   [{:keys [x y]}]
   (js/scrollTo x y))
 
-(defn restore-copy-buttons
-  "Re-initializes copy buttons from CopyButtonPlugin (highlightjs-copy - highlight.js plugin) so they look
-   correct and function correctly after the webview HTML is restored from state."
-  []
-  (.. js/document (querySelectorAll "pre code")
-      (forEach (fn [^js element]
-                 (js-delete (.. element -dataset) "highlighted")
-                 (when-let [copy-container (.. element -parentElement (querySelector ".hljs-copy-container"))]
-                   (.. copy-container (remove)))
-                 (.. js/window -hljs (highlightElement element))))))
-
 (defn handle-message
   [^js output-dom-element ^js message]
-  (let [message-data (reader/read-string (.-data message))
-        command-name (:command/name message-data)]
-    (case command-name
-      "show-result" (append-eval-result output-dom-element message-data)
-      "show-evaluated-code" (append-evaluated-code output-dom-element message-data)
-      "show-stdout" (append-stdout output-dom-element message-data)
-      "clear-output-view" (clear-output-view output-dom-element)
-      "set-code-theme" (set-code-theme! message-data)
-      "scroll-to" (scroll-to message-data)
-      "restore-copy-buttons" (restore-copy-buttons))))
+  (ensure-dom-content-loaded
+   (fn []
+     (let [message-data (reader/read-string (.-data message))
+           command-name (:command/name message-data)]
+       (case command-name
+         "show-result" (append-eval-result output-dom-element message-data)
+         "show-evaluated-code" (append-evaluated-code output-dom-element message-data)
+         "show-stdout" (append-stdout output-dom-element message-data)
+         "clear-output-view" (clear-output-view output-dom-element)
+         "set-code-theme" (set-code-theme! message-data)
+         "scroll-to" (scroll-to message-data))))))
 
 (defn handle-output-appended
   [^js _event]
   (throttled-scroll-to-bottom))
 
-(defn merge-state
-  [state]
-  (let [current-state (js->clj (.. vscode (getState)) :keywordize-keys true)
-        new-state (merge current-state state)]
-    (.. vscode (setState (clj->js new-state)))
-    ;; Send a command to the extension to save the state, so we can restore it when the webview is closed and reopened.
-    ;; TODO: Figure out why we're getting the console error `Cannot read properties of undefined (reading '__vscode_post_message__')`
-    ;; after the webview is closed and reopened.
-    ;; Every time it's closed an reopened, an additional duplicate error is added to the console.
-    ;; Note: I looked into this for a while and I'm not sure if it's worth continuing to investigate.
-    ;; It may actually be an issue with the VS Code API, but in any case, it's not causing a real problem.
-    (.. vscode (postMessage (pr-str {:command/name "save-state"
-                                     :state new-state})))))
-
-(defn save-html
-  []
-  (merge-state {:html (.. js/document.documentElement -outerHTML)}))
-
-(def throttled-save-html (throttle-fn save-html 1000))
-
-(defn handle-document-mutations
-  [_mutation-list, _observer]
-  ;; Throttle to avoid performance issues with high volume output.
-  (throttled-save-html))
-
-(defn observe-document-mutations
-  []
-  (doto (js/MutationObserver. handle-document-mutations)
-    (.observe js/document.documentElement #js {:childList true
-                                               :subtree true
-                                               :attributes true
-                                               :characterData true})))
-
-(defn save-scroll-state
-  []
-  (merge-state {:scrollLeft js/document.documentElement.scrollLeft
-                :scrollTop js/document.documentElement.scrollTop}))
-
-(def throttled-save-scroll-state (throttle-fn save-scroll-state 200))
-
 (defn add-event-listeners
   [^js output-dom-element]
   (.. js/window (addEventListener "message" (partial handle-message output-dom-element)))
-  (.. output-dom-element (addEventListener "output-appended" handle-output-appended))
-  (.. js/document (addEventListener "scroll" throttled-save-scroll-state)))
+  (.. output-dom-element (addEventListener "output-appended" handle-output-appended)))
 
 (defn ^:export main []
   (add-event-listeners output-dom-element)
-  (observe-document-mutations)
-  (.. js/window -hljs (addPlugin (CopyButtonPlugin. #js {:autohide true}))))
+  (.. hljs (registerLanguage "clojure" clojure))
+  (ensure-dom-content-loaded (fn []
+                               (.. hljs (addPlugin (CopyButtonPlugin. #js {:autohide true}))))))