+++ src/statusbar.ts
@@ -2,12 +2,14 @@ import * as vscode from 'vscode';
 import * as state from './state';
 import * as util from './utilities';
 import * as config from './config';
+import * as shadowRuntimes from './shadow-cljs-runtime';
 import { getStateValue } from '../out/cljs-lib/cljs-lib';
 import { getSession, getReplSessionTypeFromState } from './nrepl/repl-session';
 
 const connectionStatus = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1);
 const typeStatus = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1);
 const cljsBuildStatus = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1);
+const shadowRuntimeStatus = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1);
 const prettyPrintToggle = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Right, 1);
 const color = {
   active: 'white',
@@ -73,6 +75,12 @@ function update() {
   cljsBuildStatus.command = 'calva.switchCljsBuild';
   cljsBuildStatus.tooltip = undefined;
 
+  shadowRuntimeStatus.text = '';
+  shadowRuntimeStatus.command = 'calva.selectShadowCljsRuntime';
+  shadowRuntimeStatus.tooltip = undefined;
+
+  const cljsTypeName = state.extensionContext.workspaceState.get('selectedCljsTypeName');
+
   if (!getStateValue('connected')) {
     typeStatus.hide();
   }
@@ -108,6 +116,21 @@ function update() {
         cljsBuildStatus.tooltip = 'Click to connect to a CLJS build REPL';
       }
     }
+
+    if (replType === 'cljs' && cljsTypeName === 'shadow-cljs') {
+      const selectedRuntime = shadowRuntimes.getSelectedRuntimeId();
+      const runtimeInfo = shadowRuntimes.getSelectedRuntimeInfo();
+
+      if (selectedRuntime && runtimeInfo) {
+        shadowRuntimeStatus.text = `rt: ${selectedRuntime}`;
+        shadowRuntimeStatus.tooltip = `Connected to ${runtimeInfo.description}, ${runtimeInfo.sinceDescription}`;
+        shadowRuntimeStatus.command = 'calva.selectShadowCljsRuntime';
+      } else {
+        shadowRuntimeStatus.text = 'No Runtime';
+        shadowRuntimeStatus.tooltip = 'Click to select shadow-cljs runtime';
+        shadowRuntimeStatus.command = 'calva.selectShadowCljsRuntime';
+      }
+    }
     typeStatus.show();
   } else if (util.getLaunchingState()) {
     connectionStatus.color = colorValue('launchingColor', currentConf);
@@ -131,6 +154,19 @@ function update() {
   } else {
     cljsBuildStatus.hide();
   }
+
+  const replType = getReplSessionTypeFromState();
+  if (
+    getStateValue('connected') &&
+    replType === 'cljs' &&
+    cljsTypeName === 'shadow-cljs' &&
+    shadowRuntimeStatus.text
+  ) {
+    shadowRuntimeStatus.show();
+  } else {
+    shadowRuntimeStatus.hide();
+  }
+
   prettyPrintToggle.show();
 }